+++ src/imio/smartweb/core/contents/pages/footer/content.py
@@ -1,12 +1,15 @@
 # -*- coding: utf-8 -*-
 
+from imio.smartweb.common.caching import ban_physicalpath
 from imio.smartweb.core.contents import IPages
 from imio.smartweb.core.contents import Pages
 from imio.smartweb.locales import SmartwebMessageFactory as _
 from plone.autoform.directives import write_permission
 from plone.namedfile.field import NamedBlobImage
 from plone.supermodel import model
+from Products.CMFPlone.utils import parent
 from zope import schema
+from zope.globalrequest import getRequest
 from zope.interface import implementer
 
 
@@ -30,3 +33,9 @@ class IFooter(IPages):
 @implementer(IFooter)
 class Footer(Pages):
     """Footer class"""
+
+    def notifyModified(self):
+        super(Footer, self).notifyModified()
+        request = getRequest()
+        physical_path = parent(self).getPhysicalPath()
+        ban_physicalpath(request, physical_path)