+++ src/imio/smartweb/core/contents/pages/herobanner/content.py
@@ -1,11 +1,14 @@
 # -*- coding: utf-8 -*-
 
+from imio.smartweb.common.caching import ban_physicalpath
 from imio.smartweb.core.contents import IPages
 from imio.smartweb.core.contents import Pages
 from imio.smartweb.locales import SmartwebMessageFactory as _
 from plone.autoform.directives import write_permission
 from plone.supermodel import model
+from Products.CMFPlone.utils import parent
 from zope import schema
+from zope.globalrequest import getRequest
 from zope.interface import implementer
 
 
@@ -21,3 +24,9 @@ class IHeroBanner(IPages):
 @implementer(IHeroBanner)
 class HeroBanner(Pages):
     """Hero banner class"""
+
+    def notifyModified(self):
+        super(HeroBanner, self).notifyModified()
+        request = getRequest()
+        physical_path = parent(self).getPhysicalPath()
+        ban_physicalpath(request, physical_path)